Keep updated pod-groups data in a separate syncmap (#199)
