+++ pkg/scheduler/cache/status_updater/default_status_updater.go
@@ -68,6 +68,8 @@ type defaultStatusUpdater struct {
 
 	inFlightPodGroups sync.Map
 	inFlightPods      sync.Map
+
+	appliedPodGroupUpdates sync.Map
 }
 
 // +kubebuilder:rbac:groups="",resources=events,verbs=create;update;patch;delete;list;get;watch
